Fix-team-cmdlets (#168)
* Fix the handling of team project references

* Add paginator service

* Disable caching for scope properties

* Use paginator

* Add missing parameterset annotation

* Fix pipeline handling

* Improve ShouldProcess support

* Fixes collection parameter handling

* Validate release notes

* Update release notes

* Remove unused field

* Update release notes

+++ CSharp/TfsCmdlets/Controllers/Git/RemoveGitRepositoryController.cs
@@ -11,9 +11,9 @@ namespace TfsCmdlets.Controllers.Git
 
             foreach (var repo in Items)
             {
-                if (!PowerShell.ShouldProcess(Project, $"Delete Git repository '{repo.Name}'")) continue;
+                if (!PowerShell.ShouldProcess($"[Project: {repo.ProjectReference.Name}]/[Repository: {repo.Name}]", $"Delete repository")) continue;
 
-                if (!Force && !PowerShell.ShouldContinue($"Are you sure you want to delete Git repository '{repo.Name}'?")) continue;
+                if (!(repo.DefaultBranch == null || Force) && !PowerShell.ShouldContinue($"Are you sure you want to delete Git repository '{repo.Name}'?")) continue;
 
                 client.DeleteRepositoryAsync(repo.Id).Wait();
             }
